moving function/objects from wintypes into submodules
